Refactor accessibility collection into AccessibilityElement
