+++ paparazzi/src/main/java/app/cash/paparazzi/accessibility/AccessibilityOverlayDetailsView.kt
@@ -70,7 +70,7 @@ internal class AccessibilityOverlayDetailsView(context: Context) : FrameLayout(c
       canvas.drawRoundRect(badge, cornerRadius, cornerRadius, paint)
       canvas.save()
 
-      val text = it.contentDescription
+      val text = it.legendText
       val textX = badge.right + innerMargin
       val textY = badge.top
       val textLayoutWidth = (width - textX).toInt()
